+++ packages/docs/src/components/ColorsGrid.tsx
@@ -2,29 +2,29 @@ import { colors } from '@click-tools-ui/tokens'
 import { getContrast } from 'polished'
 
 export function ColorsGrid() {
-  return Object.entries(colors).map(([key, color]) => {
-    return (
-      <div
-        key={key}
-        style={{
-          background: color,
-          padding: '2rem',
-          boxShadow:
-            'rgba(0, 0, 0, 0.12) 0px 1px 3px, rgba(0, 0, 0, 0.24) 0px 1px 2px',
-        }}
-      >
-        <div
-          style={{
-            display: 'flex',
-            justifyContent: 'space-between',
-            fontFamily: 'monospace',
-            color: getContrast(color, '#FFF') < 3.5 ? '#000' : '#fff',
-          }}
-        >
-          <strong>${key}</strong>
-          <span>{color}</span>
-        </div>
-      </div>
-    )
-  })
+  return (
+    <div className="sb-unstyled divide-y-2 divide-slate-600 overflow-hidden rounded-md border-2 border-slate-600 shadow shadow-primary-900 ">
+      {Object.entries(colors).map(([key, color]) => {
+        return (
+          <div
+            key={key}
+            className="sb-unstyled p-8"
+            style={{
+              background: color,
+            }}
+          >
+            <div
+              className="flex justify-between"
+              style={{
+                color: getContrast(color, '#FFF') < 3.5 ? '#000' : '#fff',
+              }}
+            >
+              <strong>${key}</strong>
+              <span>{color}</span>
+            </div>
+          </div>
+        )
+      })}
+    </div>
+  )
 }