+++ packages/docs/src/stories/Toast.tsx
@@ -1,7 +1,0 @@
-import { FC, HTMLAttributes } from 'react'
-
-export const Toast: FC = () => {
-  return <div> Toast </div>
-}
-
-export type ToastProps = HTMLAttributes<HTMLDivElement>