+++ packages/react/tailwind.config.js
@@ -1,4 +1,10 @@
-import { colors } from '@click-tools-ui/tokens'
+import {
+  colors,
+  borderRadius,
+  fontSize,
+  lineHeight,
+  opacity,
+} from '@click-tools-ui/tokens'
 import defaultFontTheme from 'tailwindcss/defaultTheme'
 
 /** @type {import('tailwindcss').Config} */
@@ -6,11 +12,16 @@ const config = {
   content: ['./src/**/*.{js,ts,jsx,tsx}'],
   theme: {
     extend: {
-      colors,
       fontFamily: {
         sans: ['"Roboto"', ...defaultFontTheme.fontFamily.sans],
+        roboto: ['"Roboto"', ...defaultFontTheme.fontFamily.sans],
         montserrat: ['"Montserrat"', ...defaultFontTheme.fontFamily.sans],
       },
+      colors,
+      fontSize,
+      borderRadius,
+      lineHeight,
+      opacity,
     },
   },
 }