+++ packages/tokens/src/tokens/lineHeight.ts
@@ -1,0 +1,4 @@
+export const lineHeight = {
+  1.3: '1.3',
+  1.6: '1.6',
+} as const